+++ CHANGELOG.md
@@ -103,6 +103,7 @@
 
 ### Implemented enhancements
 - GP2-763 - Analytics - first sign-in
+- GP2-924 - Support thumbnails as `poster` images for <video> nodes
 - GP2-748 - Analytics product classifier - match found
 - GP2-689 - Analytics product classifier - product selected
 - GP2-690 - Analytics target market open